Objects -> primitives

+++ src/test/java/no/stelar7/api/l4j8/tests/summoner/SummonerMasteriesTest.java
@@ -18,24 +18,24 @@ public class SummonerMasteriesTest
     private final Consumer<List<MasteryPage>> doAssertions = (List<MasteryPage> pages) ->
     {
         pages.forEach(page ->
-                                 {
+                      {
             
-                                     Assert.assertNotNull("SpectatorMastery Page does not have an id", page.getId());
-                                     // Assert.assertNotNull("SpectatorMastery Page does not have a name",
-                                     // page.getName()); Names can be empty, so ignore this...
-                                     Assert.assertNotNull("SpectatorMastery Page does not contain any masteries", page.getMasteries());
-                                     Assert.assertNotNull("Unable to determine current SpectatorMastery page", page.isCurrent());
+                          Assert.assertNotNull("SpectatorMastery Page does not have an id", page.getId());
+                          // Assert.assertNotNull("SpectatorMastery Page does not have a name",
+                          // page.getName()); Names can be empty, so ignore this...
+                          Assert.assertNotNull("SpectatorMastery Page does not contain any masteries", page.getMasteries());
+                          Assert.assertNotNull("Unable to determine current SpectatorMastery page", page.isCurrent());
             
-                                     page.getMasteries().forEach((final Mastery mastery) ->
-                                                                 {
-                                                                     Assert.assertNotNull("SpectatorMastery does not have an id", mastery.getId());
-                                                                     Assert.assertNotNull("SpectatorMastery does not have a rank", mastery.getRank());
+                          page.getMasteries().forEach((final Mastery mastery) ->
+                                                      {
+                                                          Assert.assertNotNull("SpectatorMastery does not have an id", mastery.getId());
+                                                          Assert.assertNotNull("SpectatorMastery does not have a rank", mastery.getRank());
                 
-                                                                     Assert.assertNotEquals("SpectatorMastery does not have a valid id", mastery.getId(), (Integer) 0);
-                                                                     Assert.assertNotEquals("SpectatorMastery does not have a valid rank", mastery.getRank(), (Integer) 0);
-                                                                 });
+                                                          Assert.assertNotEquals("SpectatorMastery does not have a valid id", mastery.getId(), 0);
+                                                          Assert.assertNotEquals("SpectatorMastery does not have a valid rank", mastery.getRank(), 0);
+                                                      });
             
-                                 });
+                      });
     };
     
     @Test
